requirements.txt
numpy
pillow
albumentations>=0.2.2
opencv-python-headless
torch>=1.4.0
torchvision>=0.5.0
tqdm>=4.29.0
requests
geojson
mercantile>=1.0.4
osmium>=2.15.0
rasterio>=1.1.1
supermercado>=0.0.5
shapely>=1.6.4
pyproj>=1.9.6
toml
webcolors
psycopg2-binary
